Prof. Ozay Squashes Bug + Measuring Closeness of Analytical Model
Prof. Ozay ID'ed that the velocity (or mu) variable starts at 0 which
would explain why the system quickly produced NaNs.

Next, I'm working on simulating the system evolution but with a
"correction" every N steps so that at every Nth sample, we start the
simulated trajectory on the measured trajectory and then observe how
long it takes to diverge.