Control which values u[t-1] when selecting u[t]
The test mcity_data_practice2.m introduces a controller design algorithm
which should make the car's motion smoother. It introduces a term in the
Quadratic Program which essentially places preference on control inputs
u[t] which are close to the previous input.

Also introduced the mcity_data_analysis() function which is used to run
each experiment that I have been using while we try to iron out
bugs\non-smoothness in the current controller.

The lk_pcis_controller.m file (defining the lk_pcis_controller) was
modified so that:
1. One of the outputs of the obj.step() function is well-defined.
(qp_status now accurately describes the status that was given by
mpcqpsolver).
2. Variables in the class that are related to the optimization cost are
available to an external user (this is necessary to introduce the new
control preference/smoothing).